fix warning in scicos MVC compilation

+++ scilab/modules/scicos/src/cpp/view_scilab/BaseAdapter.hxx
@@ -144,7 +144,7 @@ public:
         std::sort(properties.begin(), properties.end(), property<Adaptor>::original_index_cmp);
 
         // create the header
-        types::String* header = new types::String(1 + properties.size(), 1);
+        types::String* header = new types::String(1 + (int)properties.size(), 1);
         header->set(0, Adaptor::getSharedTypeStr().c_str());
         int index = 1;
         for (typename property<Adaptor>::props_t_it it = properties.begin(); it != properties.end(); ++it, ++index)
